requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
EditorConfig==0.12.1
flake8==3.3.0
flake8-mypy==17.3.3
greenlet==0.4.12
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
jsbeautifier==1.6.14
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
mypy==0.511
neovim==0.1.13
pbr==3.1.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
pyOpenSSL==17.1.0
requests==2.18.1
service-identity==17.0.0
six==1.10.0
treq==17.3.1
Twisted==17.5.0
typed-ast==1.0.4
urllib3==1.21.1
zope.interface==4.4.2
pytz
